Remove tsx/jsx/js and make typescript use tsx
