+++ bindings/python/Cargo.toml
@@ -28,9 +28,6 @@ log = { workspace = true }
 [features]
 python = [ "codegen-sdk-analyzer/python"]
 typescript = [ "codegen-sdk-analyzer/typescript"]
-tsx = [ "codegen-sdk-analyzer/tsx"]
-jsx = [ "codegen-sdk-analyzer/jsx"]
-javascript = [ "codegen-sdk-analyzer/javascript"]
 json = [ "codegen-sdk-analyzer/json"]
 java = [ "codegen-sdk-analyzer/java"]
 ruby = [ "codegen-sdk-analyzer/ruby"]
@@ -39,5 +36,5 @@ go = [ "codegen-sdk-analyzer/go"]
 markdown = [ "codegen-sdk-analyzer/markdown"]
 yaml = [ "codegen-sdk-analyzer/yaml"]
 toml = [ "codegen-sdk-analyzer/toml"]
-stable = ["codegen-sdk-analyzer/stable", "typescript", "tsx", "jsx", "javascript", "json", "toml", "yaml", "java", "ruby", "rust", "go", "markdown"]
+stable = ["codegen-sdk-analyzer/stable", "typescript", "json", "toml", "yaml", "java", "ruby", "rust", "go", "markdown"]
 default = ["python", "typescript"]
