--- codegen-sdk-common/src/language/jsx.rs
@@ -1,12 +1,0 @@
-use super::Language;
-lazy_static! {
-    pub static ref JSX: Language = Language::new(
-        "jsx",
-        "JSX",
-        tree_sitter_typescript::TSX_NODE_TYPES,
-        &["jsx"],
-        tree_sitter_typescript::LANGUAGE_TSX.into(),
-        tree_sitter_typescript::TAGS_QUERY,
-    )
-    .unwrap();
-}
