--- languages/codegen-sdk-tsx/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-version.workspace = true
-edition.workspace = true
-readme.workspace = true
-license.workspace = true
-homepage.workspace = true
-authors.workspace = true
-repository.workspace = true
-name = "codegen-sdk-tsx"
-
-description = "TSX language support for codegen-sdk"
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-
-[dependencies]
-codegen-sdk-common = { workspace = true, features = ["typescript"] }
-salsa = { workspace = true }
-tree-sitter = { workspace = true }
-derive_more = { workspace = true }
-ambassador = { workspace = true }
-indextree = { workspace = true }
-subenum = {workspace = true}
-bytes = { workspace = true }
-codegen-sdk-cst = { workspace = true }
-log = { workspace = true }
-codegen-sdk-ast = { workspace = true }
-codegen-sdk-resolution = { workspace = true }
-[build-dependencies]
-codegen-sdk-cst-generator = { workspace = true }
-codegen-sdk-ast-generator = { workspace = true }
-codegen-sdk-common = { workspace = true, features = ["typescript"] }
-env_logger = { workspace = true }
-log = { workspace = true }
-[features]
-serialization = ["codegen-sdk-common/serialization"]
